Close #5 - Fix some user fields raising exception
If a form is created with a certain number of columns, but the record
data tries to add data for a column that doesn't exist, don't add those
columns. Only add the system columns that we know should exist.

We can't look up the column names due to a limitation with
fusion_tables.

This meant refactoring RecordData as well as the record classes to be
able to try using the fusion formatted data first and if that fails,
then trying with the regular, raw record data.

Add convenience methods to RecordProcessor::Base as well.