Collapse sections when looking up form fields
We won't ever receive a section in record data, but we will receive
data for the sections contained therein.
Collapse the section by replacing it with its inner elements so we can
get at the inner fields.
Update the specs to include this example.
Also make the specs more readable by using describe blocks.