Pass form in to RecordData and extractor
Allows us to make sure the fulcrum form exists before we even call this
at all.
This is important in the case we get an event and the fulcrum form
doesn’t exist at all.
Update the specs where it is used to pass in the form.
Add a context to provide a default form for use in the specs.